Refactoring to ECMAScript 6 and add Babel JS
