+++ babel.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  presets: [
+    '@babel/preset-env',
+  ],
+  targets: {
+    node: 'current',
+  },
+};
