+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  transform: {
+    '^.+\\.[t|j]sx?$': 'babel-jest',
+  },
+};
