+++ src/bin/index.js
@@ -1,0 +1,8 @@
+#! /usr/bin/env node
+
+/* eslint-disable node/no-unpublished-import */
+/*jshint esversion: 8 */
+
+import { passwordGenerator } from '../lib/password-generator.js';
+
+passwordGenerator();
