+++ src/dictionaries/adjectives.json
