+++ src/dictionaries/adverbs.json
