+++ src/dictionaries/animals.json
