+++ src/dictionaries/emoji.json
