+++ src/dictionaries/fruits.json
