+++ src/dictionaries/gemstones.json
