+++ src/dictionaries/hazards.json
