+++ src/dictionaries/instruments.json
