+++ src/dictionaries/lovecraft.json
