+++ src/dictionaries/metals.json
