+++ src/dictionaries/shakespeare.json
