+++ src/dictionaries/strange.json
