+++ src/utils/randomSyllable.js
@@ -1,8 +1,8 @@
 /* eslint-disable node/no-unpublished-import */
 /*jshint esversion: 8 */
 
-import { randomConsonant } from "./randomConsonant.js";
-import { randomVowel } from "./randomVowel.js";
+import { randomConsonant } from "../../src/utils/randomConsonant.js";
+import { randomVowel } from "../../src/utils/randomVowel.js";
 
 export function randomSyllable() {
   return randomConsonant() + randomVowel() + randomConsonant();
