+++ src/utils/randomVowel.js
@@ -1,7 +1,7 @@
 /* eslint-disable node/no-unpublished-import */
 /*jshint esversion: 8 */
 
-import { randomNumber } from "./randomNumber.js";
+import { randomNumber } from "../../src/utils/randomNumber.js";
 
 // There are 5 vowels (a, e, i, o, u)
 let vowels = "aeiou";
