+++ tests/utils/toSnakeCase.test.js
@@ -1,6 +1,6 @@
 /*jshint esversion: 8 */
 
-import { toSnakeCase } from "../../utils/toSnakeCase.js";
+import { toSnakeCase } from "../../src/utils/toSnakeCase.js";
 import * as chai from "chai";
 let expect = chai.expect;
